chore(test): refactor debugger integration tests
Split up the very large basic.spec.js file into smaller scoped files
