+++ integration-tests/debugger/utils.js
@@ -1,14 +1,17 @@
 'use strict'
 
 const assert = require('assert')
+const os = require('os')
 const { basename, join } = require('path')
 const { readFileSync } = require('fs')
 const { randomUUID } = require('crypto')
 
 const Axios = require('axios')
 
+const { assertObjectContains, assertUUID } = require('../helpers')
 const { sandboxCwd, useSandbox, FakeAgent, spawnProc } = require('../helpers')
 const { generateProbeConfig } = require('../../packages/dd-trace/test/debugger/devtools_client/utils')
+const { version } = require('../../package.json')
 
 const BREAKPOINT_TOKEN = '// BREAKPOINT'
 const pollInterval = 0.1
@@ -70,8 +73,13 @@ const pollInterval = 0.1
  */
 
 module.exports = {
+  assertBasicInputPayload,
   pollInterval,
-  setup
+  setup,
+  setupAssertionListeners,
+  testBasicInput,
+  testBasicInputWithoutDD,
+  testBasicInputWithoutRC,
 }
 
 /**
@@ -256,4 +264,153 @@ function getBreakpointInfo ({ deployedFile, sourceFile = deployedFile, stackInde
   }
 
   return result
+}
+
+/**
+ * Test helper for basic input messages with remote config and tracing.
+ *
+ * @param {DebuggerTestEnvironment} t - The test environment.
+ * @param {Function} done - The mocha done callback.
+ */
+function testBasicInput (t, done) {
+  t.triggerBreakpoint()
+  setupAssertionListeners(t, done)
+  t.agent.addRemoteConfig(t.rcConfig)
+}
+
+/**
+ * Test helper for basic input messages without remote config (e.g., from probe file).
+ *
+ * @param {DebuggerTestEnvironment} t - The test environment.
+ * @param {import('../../packages/dd-trace/test/debugger/devtools_client/utils').ProbeConfig} probe - The probe config.
+ * @param {Function} done - The mocha done callback.
+ */
+function testBasicInputWithoutRC (t, probe, done) {
+  t.triggerBreakpoint()
+  setupAssertionListeners(t, done, probe)
+}
+
+/**
+ * Setup assertion listeners for basic input tests with tracing integration.
+ *
+ * @param {DebuggerTestEnvironment} t - The test environment.
+ * @param {Function} done - The mocha done callback.
+ * @param {import('../../packages/dd-trace/test/debugger/devtools_client/utils').ProbeConfig} [probe] - Optional probe
+ *   config to use instead of t.rcConfig.config.
+ */
+function setupAssertionListeners (t, done, probe) {
+  let traceId, spanId, dd
+
+  t.agent.on('message', ({ payload }) => {
+    const span = payload.find((arr) => arr[0].name === 'fastify.request')?.[0]
+    if (!span) return
+
+    traceId = span.trace_id.toString()
+    spanId = span.span_id.toString()
+
+    assertDD()
+  })
+
+  t.agent.on('debugger-input', ({ payload }) => {
+    assertBasicInputPayload(t, payload, probe)
+
+    payload = payload[0]
+    assert.ok(typeof payload.dd === 'object' && payload.dd !== null)
+    assert.deepStrictEqual(['span_id', 'trace_id'], Object.keys(payload.dd).sort())
+    assert.strictEqual(typeof payload.dd.trace_id, 'string')
+    assert.strictEqual(typeof payload.dd.span_id, 'string')
+    assert.ok(payload.dd.trace_id.length > 0)
+    assert.ok(payload.dd.span_id.length > 0)
+    dd = payload.dd
+
+    assertDD()
+  })
+
+  function assertDD () {
+    if (!traceId || !spanId || !dd) return
+    assert.strictEqual(dd.trace_id, traceId)
+    assert.strictEqual(dd.span_id, spanId)
+    done()
+  }
+}
+
+/**
+ * Test helper for basic input messages without DD tracing integration.
+ *
+ * @param {DebuggerTestEnvironment} t - The test environment.
+ * @param {Function} done - The mocha done callback.
+ */
+function testBasicInputWithoutDD (t, done) {
+  t.triggerBreakpoint()
+
+  t.agent.on('debugger-input', ({ payload }) => {
+    assertBasicInputPayload(t, payload)
+    assert.ok(!('dd' in payload[0]))
+    done()
+  })
+
+  t.agent.addRemoteConfig(t.rcConfig)
+}
+
+/**
+ * Assert that the basic input payload structure and content is correct.
+ *
+ * @param {DebuggerTestEnvironment} t - The test environment.
+ * @param {Array<object>} payload - The debugger input payload.
+ * @param {import('../../packages/dd-trace/test/debugger/devtools_client/utils').ProbeConfig} [probe] - Optional probe
+ *   config to use instead of t.rcConfig.config.
+ */
+function assertBasicInputPayload (t, payload, probe = t.rcConfig.config) {
+  assert.ok(Array.isArray(payload))
+  assert.strictEqual(payload.length, 1)
+  const data = payload[0]
+
+  const expected = {
+    ddsource: 'dd_debugger',
+    hostname: os.hostname(),
+    service: 'node',
+    message: 'Hello World!',
+    logger: {
+      name: t.breakpoint.deployedFile,
+      method: 'fooHandler',
+      version,
+      thread_name: 'MainThread'
+    },
+    debugger: {
+      snapshot: {
+        probe: {
+          id: probe.id,
+          version: 0,
+          location: { file: t.breakpoint.deployedFile, lines: [String(t.breakpoint.line)] }
+        },
+        language: 'javascript'
+      }
+    }
+  }
+
+  assertObjectContains(data, expected)
+
+  assert.match(data.logger.thread_id, /^pid:\d+$/)
+
+  assertUUID(data.debugger.snapshot.id)
+  assert.strictEqual(typeof data.debugger.snapshot.timestamp, 'number')
+  assert.ok(data.debugger.snapshot.timestamp > Date.now() - 1000 * 60)
+  assert.ok(data.debugger.snapshot.timestamp <= Date.now())
+
+  assert.ok(Array.isArray(data.debugger.snapshot.stack))
+  assert.ok(data.debugger.snapshot.stack.length > 0)
+  for (const frame of data.debugger.snapshot.stack) {
+    assert.ok(typeof frame === 'object' && frame !== null)
+    assert.deepStrictEqual(['columnNumber', 'fileName', 'function', 'lineNumber'], Object.keys(frame).sort())
+    assert.strictEqual(typeof frame.fileName, 'string')
+    assert.strictEqual(typeof frame.function, 'string')
+    assert.ok(frame.lineNumber > 0)
+    assert.ok(frame.columnNumber > 0)
+  }
+  const topFrame = data.debugger.snapshot.stack[0]
+  // path seems to be prefixed with `/private` on Mac
+  assert.match(topFrame.fileName, new RegExp(`${t.appFile}$`))
+  assert.strictEqual(topFrame.function, 'fooHandler')
+  assert.strictEqual(topFrame.lineNumber, t.breakpoint.line)
+  assert.strictEqual(topFrame.columnNumber, 3)
 }
